src/api/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.9.1
azure-ai-agents==1.1.0
azure-ai-projects>=2.0.0b1
azure-identity
azure-common==1.1.28
azure-core==1.35.0
azure-cosmos==4.5.1
azure-functions==1.24.0
azure-identity==1.25.0
azure-keyvault-secrets==4.4.0
azure-monitor-opentelemetry-exporter==1.0.0b19
azure-storage-blob==12.26.0
beanie==1.11.12
beautifulsoup4==4.14.2
cachetools==6.2.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
coverage==7.11.0
cryptography==45.0.6
Deprecated==1.2.18
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.95.2
fixedint==0.1.6
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.25.2
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jwcrypto==1.5.6
lxml==6.0.2
MarkupSafe==3.0.3
motor==3.7.1
msal==1.33.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.6.4
oauthlib==3.3.1
openai==1.108.2
opentelemetry-api==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
orjson==3.11.3
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.22
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.14.1
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dotenv==0.20.0
python-multipart==0.0.9
PyYAML==6.0.3
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.27.0
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.19.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
